DONE: Instructor Edit

+++ lib/pages/super_admin/edit_pages/edit_instructor_row_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:sti_startnow/main.dart';
 import 'package:sti_startnow/pages/components/buttons/bottom_button.dart';
 import 'package:sti_startnow/pages/components/buttons/delete_button.dart';
+import 'package:sti_startnow/pages/components/custom_bottom_sheet.dart';
 import 'package:sti_startnow/pages/components/page_app_bar.dart';
 import 'package:sti_startnow/pages/components/text_input.dart';
 import 'package:sti_startnow/theme/app_theme.dart';
@@ -123,36 +124,46 @@ class _EditInstructorRowPageState extends State<EditInstructorRowPage> {
                 child: Column(
                   children: [
                     BottomButton(
-                      onPressed: () async {
-                        if (_formKey.currentState!.validate()) {
-                          // Show circular progress indicator
-                          showDialog(
-                            context: context,
-                            builder: (context) {
-                              return PopScope(
-                                canPop: false,
-                                child: Center(
-                                  child: const CircularProgressIndicator(),
-                                ),
-                              );
-                            },
-                          );
+                      onPressed: () {
+                        showModalBottomSheet(
+                          context: context, 
+                          builder: (context) {
+                            return CustomBottomSheet(
+                              subtitle: "Changing the information at the database.",
+                              submitFunc: () async {
+                                if (_formKey.currentState!.validate()) {
+                                  // Show circular progress indicator
+                                  showDialog(
+                                    context: context,
+                                    builder: (context) {
+                                      return PopScope(
+                                        canPop: false,
+                                        child: Center(
+                                          child: const CircularProgressIndicator(),
+                                        ),
+                                      );
+                                    },
+                                  );
 
-                          await supabase
-                              .from("PROFESSOR")
-                              .update({
-                                'prof_fname': firstNameController.text,
-                                'prof_lname': lastNameController.text,
-                                'department': departmentController.text,
-                                'email': emailAddressController.text,
-                              })
-                              .eq('prof_id', profID);
+                                  await supabase
+                                      .from("PROFESSOR")
+                                      .update({
+                                        'prof_fname': firstNameController.text,
+                                        'prof_lname': lastNameController.text,
+                                        'department': departmentController.text,
+                                        'email': emailAddressController.text,
+                                      })
+                                      .eq('prof_id', profID);
 
-                          if (context.mounted) {
-                            Navigator.pop(context);
-                            Navigator.pop(context);
+                                  if (context.mounted) {
+                                    Navigator.pop(context);
+                                    Navigator.pop(context);
+                                  }
+                                }
+                              },
+                            );
                           }
-                        }
+                        );
                       },
                       text: "Save",
                     ),
@@ -160,29 +171,39 @@ class _EditInstructorRowPageState extends State<EditInstructorRowPage> {
                     const SizedBox(height: 10),
 
                     DeleteButton(
-                      onPressed: () async {
-                        // Show circular progress indicator
-                        showDialog(
-                          context: context,
+                      onPressed: () {
+                        showModalBottomSheet(
+                          context: context, 
                           builder: (context) {
-                            return PopScope(
-                              canPop: false,
-                              child: Center(
-                                child: const CircularProgressIndicator(),
-                              ),
+                            return CustomBottomSheet(
+                              subtitle: "Permanently deleting a record.",
+                              submitFunc: () async {
+                                // Show circular progress indicator
+                                showDialog(
+                                  context: context,
+                                  builder: (context) {
+                                    return PopScope(
+                                      canPop: false,
+                                      child: Center(
+                                        child: const CircularProgressIndicator(),
+                                      ),
+                                    );
+                                  },
+                                );
+
+                                await supabase
+                                    .from("PROFESSOR")
+                                    .delete()
+                                    .eq('prof_id', profID);
+
+                                if (context.mounted) {
+                                  Navigator.pop(context);
+                                  Navigator.pop(context);
+                                }
+                              },
                             );
-                          },
+                          }
                         );
-
-                        await supabase
-                            .from("PROFESSOR")
-                            .delete()
-                            .eq('prof_id', profID);
-
-                        if (context.mounted) {
-                          Navigator.pop(context);
-                          Navigator.pop(context);
-                        }
                       },
                       text: "Delete",
                     ),
